Keep SAM header chrom order same as reference
Previously, chromosomes were alphabetically sorted when writing the SAM
header. Generally, this isn't an issue if you're using a standard
reference from ENSEMBL or GENCODE. However, when a custom reference with
chromosomes that aren't in alphabetical order is used in biscuit align,
the order will differ between the SAM/BAM and the reference.

Further, BISCUIT wasn't handling chromosome names correctly if the
sorted order differed from the reference order. Therefore, to fix the
mapping and bring the reference and SAM chromosome order into alignment,
chromosome names will not be alphabetically sorted going forward.