make biscuit sam header match bwa
biscuit didn't output an @HD SAM tag in its header, while bwa does. The
@HD tag isn't required; however, it is helpful in determining the sort
order and how reads are grouped. Following bwa's convention for the
header allows biscuit to output this information as well.

It also adds the AH value within the @SQ tags for indicating which reads
are considered alternate loci.