Resolve #65
VERY weird bug. When running example data, one read would shift the
pointer to a config struct by 1 byte in memory which would ultimately
cause a string comparison against garbage and give a seg fault. I'm not
entirely sure why this would happen. To the best of my knowledge, it
wasn't due to an overflow in trying to write out of bounds for an
array, but increasing the array size for the max cinread read length by
one (from 301 to 302) resolved this issue.

This seems like such a small corner case that I'm making note that this
happens, but am going to be happy with the apparent band-aid fix of
adjusting array sizes fixing the issue.